feat(storage): TableOption for common
