+++ src/meta/src/hummock/compaction_group/manager.rs
@@ -16,13 +16,14 @@ use std::collections::{BTreeMap, HashMap, HashSet};
 use std::sync::Arc;
 
 use itertools::Itertools;
+use risingwave_common::catalog::TableOption;
 use risingwave_hummock_sdk::compaction_group::{Prefix, StaticCompactionGroupId};
 use risingwave_hummock_sdk::CompactionGroupId;
 use risingwave_pb::hummock::CompactionConfig;
 use tokio::sync::RwLock;
 
 use crate::hummock::compaction::compaction_config::CompactionConfigBuilder;
-use crate::hummock::compaction_group::{CompactionGroup, TableOption};
+use crate::hummock::compaction_group::CompactionGroup;
 use crate::hummock::error::{Error, Result};
 use crate::manager::{MetaSrvEnv, SourceId};
 use crate::model::{MetadataModel, TableFragments, ValTransaction, VarTransaction};
@@ -81,7 +82,7 @@ impl<S: MetaStore> CompactionGroupManager<S> {
         table_fragments: &TableFragments,
         table_properties: &HashMap<String, String>,
     ) -> Result<()> {
-        let table_option = CompactionGroup::build_table_option(table_properties);
+        let table_option = TableOption::build_table_option(table_properties);
         let mut pairs = vec![];
         // materialized_view or materialized_source
         pairs.push((
@@ -129,7 +130,7 @@ impl<S: MetaStore> CompactionGroupManager<S> {
         source_id: u32,
         table_properties: &HashMap<String, String>,
     ) -> Result<()> {
-        let table_option = CompactionGroup::build_table_option(table_properties);
+        let table_option = TableOption::build_table_option(table_properties);
         self.inner
             .write()
             .await
@@ -350,11 +351,11 @@ mod tests {
     use std::collections::HashMap;
     use std::ops::Deref;
 
-    use risingwave_common::catalog::TableId;
+    use risingwave_common::catalog::{TableId, TableOption};
     use risingwave_hummock_sdk::compaction_group::{Prefix, StaticCompactionGroupId};
 
     use crate::hummock::compaction_group::manager::{
-        CompactionGroup, CompactionGroupManager, CompactionGroupManagerInner,
+        CompactionGroupManager, CompactionGroupManagerInner,
     };
     use crate::hummock::test_utils::setup_compute_env;
     use crate::model::TableFragments;
@@ -385,7 +386,7 @@ mod tests {
         assert_eq!(registered_number(inner.read().await.deref()), 0);
 
         let table_properties = HashMap::from([(String::from("ttl"), String::from("300"))]);
-        let table_option = CompactionGroup::build_table_option(&table_properties);
+        let table_option = TableOption::build_table_option(&table_properties);
 
         // Test register
         inner
